refactor: reduce panels agent health baseline
