+++ src/main/ipc/handlers/panels.ts
@@ -43,207 +43,15 @@ import {
   generateBedContent
 } from './panels-logic'
 import type { PanelCacheCallbacks } from './panels-logic'
-import type { StorageSession } from '../../storage/session'
-import type { GeneReferenceDb } from '../../database/GeneReferenceDb'
-import type { CreatePanelInput, PanelGeneRow, PanelRow } from '../../../shared/types/panels'
-import type { PanelAppPanel } from '../../services/api/PanelAppClient'
+import {
+  importPanelAppForSession,
+  generateStringDbForSession,
+  generateBedContentForSession
+} from './panels-session'
 
 /** Shared cache-clearing callback wired to the variant module's interval cache. */
 const cacheCallbacks: PanelCacheCallbacks = {
   clearPanelIntervalCache: () => clearPanelIntervalCache()
-}
-
-const GREEN_LEVELS = new Set(['3', '4', 'green'])
-const GREEN_AMBER_LEVELS = new Set(['2', '3', '4', 'green', 'amber'])
-
-interface ResolvedGene {
-  hgncId: string
-  symbol: string
-}
-
-interface ValidationResult {
-  status: string
-  hgncId?: string
-  symbol?: string
-  currentSymbol?: string
-}
-
-function resolveValidatedGenes(validationResults: ValidationResult[]): ResolvedGene[] {
-  const resolved: ResolvedGene[] = []
-  const seenHgnc = new Set<string>()
-
-  for (const result of validationResults) {
-    const hgncId = result.hgncId
-    const symbol =
-      result.status === 'approved'
-        ? result.symbol
-        : result.status === 'alias'
-          ? result.currentSymbol
-          : undefined
-
-    if (hgncId !== undefined && symbol !== undefined && !seenHgnc.has(hgncId)) {
-      seenHgnc.add(hgncId)
-      resolved.push({ hgncId, symbol })
-    }
-  }
-
-  return resolved
-}
-
-async function createPanelWithGenes(
-  session: StorageSession,
-  input: CreatePanelInput,
-  genes: ResolvedGene[]
-): Promise<PanelRow & { genes: PanelGeneRow[] }> {
-  const createdPanel = (await session
-    .getWriteExecutor()
-    .execute({ type: 'panels:create', params: [input] })) as PanelRow
-
-  if (genes.length > 0) {
-    await session
-      .getWriteExecutor()
-      .execute({ type: 'panels:setGenes', params: [createdPanel.id, genes] })
-  }
-
-  const panelGenes = (await session
-    .getReadExecutor()
-    .execute({ type: 'panels:getGenes', params: [createdPanel.id] })) as PanelGeneRow[]
-
-  cacheCallbacks.clearPanelIntervalCache()
-  return { ...createdPanel, genes: panelGenes }
-}
-
-async function importPanelAppForSession(
-  session: StorageSession,
-  params: {
-    panelId: number
-    region: 'uk' | 'aus'
-    confidenceThreshold: string
-    name?: string
-  },
-  geneRef: GeneReferenceDb,
-  client: PanelAppClient
-): Promise<PanelRow & { genes: PanelGeneRow[] }> {
-  const panel: PanelAppPanel = await client.getPanel(params.panelId, params.region)
-  const confidenceSet =
-    params.confidenceThreshold === 'green'
-      ? GREEN_LEVELS
-      : params.confidenceThreshold === 'green_amber'
-        ? GREEN_AMBER_LEVELS
-        : null
-  const filteredGenes =
-    confidenceSet === null
-      ? panel.genes
-      : panel.genes.filter((gene) => confidenceSet.has(gene.confidence_level))
-  const symbols = filteredGenes.map((gene) => gene.gene_data.gene_symbol)
-  const resolvedGenes = resolveValidatedGenes(geneRef.validateSymbols(symbols))
-  const source = params.region === 'uk' ? 'panelapp_uk' : 'panelapp_aus'
-
-  return createPanelWithGenes(
-    session,
-    {
-      name: params.name ?? `${panel.name} (PanelApp ${params.region.toUpperCase()})`,
-      description: `Imported from PanelApp ${params.region.toUpperCase()} v${panel.version}`,
-      version: panel.version,
-      source,
-      sourceId: String(params.panelId),
-      sourceMetadata: {
-        confidence_threshold: params.confidenceThreshold,
-        total_genes: panel.genes.length,
-        filtered_genes: filteredGenes.length,
-        resolved_genes: resolvedGenes.length,
-        panel_version: panel.version
-      }
-    },
-    resolvedGenes
-  )
-}
-
-async function generateStringDbForSession(
-  session: StorageSession,
-  params: {
-    seedGenes: string[]
-    requiredScore: number
-    networkType: 'physical' | 'functional'
-    name?: string
-  },
-  geneRef: GeneReferenceDb,
-  client: StringDbClient
-): Promise<PanelRow & { genes: PanelGeneRow[] }> {
-  const partners = await client.getInteractionPartners(params.seedGenes, {
-    requiredScore: params.requiredScore,
-    networkType: params.networkType
-  })
-  const allSymbols = [...params.seedGenes, ...partners.map((partner) => partner.symbol)]
-  const resolvedGenes = resolveValidatedGenes(geneRef.validateSymbols(allSymbols))
-
-  return createPanelWithGenes(
-    session,
-    {
-      name:
-        params.name ??
-        `StringDB Network (${params.seedGenes.slice(0, 3).join(', ')}${
-          params.seedGenes.length > 3 ? '...' : ''
-        })`,
-      description: `Generated from StringDB ${params.networkType} network (score >= ${params.requiredScore})`,
-      source: 'stringdb',
-      sourceMetadata: {
-        seed_genes: params.seedGenes,
-        score_threshold: params.requiredScore,
-        network_type: params.networkType,
-        partners_found: partners.length
-      }
-    },
-    resolvedGenes
-  )
-}
-
-async function generateBedContentForSession(
-  session: StorageSession,
-  panelId: number,
-  assembly: string,
-  paddingBp: number,
-  geneRef: GeneReferenceDb
-): Promise<{ lines: string[]; panelName: string; geneCount: number }> {
-  const panel = (await session.getReadExecutor().execute({
-    type: 'panels:get',
-    params: [panelId]
-  })) as PanelRow | null
-  if (panel === null) throw new Error(`Panel ${panelId} not found`)
-
-  const genes = (await session.getReadExecutor().execute({
-    type: 'panels:getGenes',
-    params: [panelId]
-  })) as PanelGeneRow[]
-  if (genes.length === 0) {
-    throw new Error('Panel has no genes to export')
-  }
-
-  const coordsMap = geneRef.getCoordinatesForGenes(
-    genes.map((gene) => gene.hgnc_id),
-    assembly
-  )
-  if (coordsMap.size === 0) {
-    throw new Error(`No coordinates found for assembly ${assembly}`)
-  }
-
-  const lines = [`track name="${panel.name}" description="Gene panel: ${panel.name}"`]
-  for (const gene of genes) {
-    const coords = coordsMap.get(gene.hgnc_id)
-    if (coords === undefined) continue
-
-    const chr = coords.chromosome.startsWith('chr') ? coords.chromosome : `chr${coords.chromosome}`
-    const bedStart = Math.max(0, coords.start_pos - 1 - paddingBp)
-    const bedEnd = coords.end_pos + paddingBp
-    lines.push(`${chr}\t${bedStart}\t${bedEnd}\t${gene.symbol}`)
-  }
-
-  mainLogger.info(
-    `Generated BED content for panel "${panel.name}" (${coordsMap.size} genes)`,
-    'panels'
-  )
-
-  return { lines, panelName: panel.name, geneCount: coordsMap.size }
 }
 
 /**
@@ -538,7 +346,7 @@ export function registerPanelHandlers({ ipcMain, getDb, getDbManager }: HandlerD
       const geneRef = getGeneReferenceDb()
       const session = getDbManager().getCurrentSession()
       if (session.capabilities.backend === 'postgres') {
-        return importPanelAppForSession(session, validated.data, geneRef, client)
+        return importPanelAppForSession(session, validated.data, geneRef, client, cacheCallbacks)
       }
       return importPanelApp(validated.data, getDb, geneRef, client, cacheCallbacks)
     })
@@ -558,7 +366,7 @@ export function registerPanelHandlers({ ipcMain, getDb, getDbManager }: HandlerD
       const geneRef = getGeneReferenceDb()
       const session = getDbManager().getCurrentSession()
       if (session.capabilities.backend === 'postgres') {
-        return generateStringDbForSession(session, validated.data, geneRef, client)
+        return generateStringDbForSession(session, validated.data, geneRef, client, cacheCallbacks)
       }
       return generateStringDb(validated.data, getDb, geneRef, client, cacheCallbacks)
     })
